requirements.txt
numpy
pandas
geopandas
python-dotenv